requirements.txt
django==4.1.3
djangorestframework==3.14.0
django-auto-prefetching==0.2.12
django-utils-six==2.0
django-cors-headers==3.14.0
drf-yasg==1.21.5
ethereum==2.3.2
pyqrcode==1.2.1
requests==2.28.1
web3==5.30.0
bit==0.8.0
python-dotenv==1.0.0